Fix longstanding bug in issue #31.
The issue describe here

https://github.com/credativ/informix_fdw/issues/31

uncovered an unfortunate mistake in our approach on how
we plan a modify action. Formerly, ifxPlanForeignModify() was
also responsible to create and execute the necessary query objects
on the remote Informix server (like describing the statement and thus
creating the required SQLDA structure).

In this issue above it turned out that we are far too optimistic in
doing this preparation work at this early stage: ifxPlanForeignModify() might not
be called once a prepared statement settles the execution plan and thus
we miss important things like creating and describing the statement handles.

To fix this, we need to delay creating of query objects on the remote
Informix server to ifxBeginForeignModify(), which is now responsible
to create and describe the statement.

ifxPlanForeignModify() still creates all required identifiers for
query objects (like cursor, statement handle, ...), but this leaves
us after the plan settled with a static reference counter (refid) used
in the identifiers.
I think this isn't a problem here, since afaics the refid can't be reused
somewhere else at this point. This might be wrong, but since we pass
regression tests at this point i leave it for now.