Follow up fix for issue https://github.com/credativ/informix_fdw/issues/31.
Commit 285119f024a0137ba84a3fc7b825b9e2ccf107bf partially fixed our
architecture thinko regarding query planning, but it turned out
that we have the same issue in ifxGetForeignRelSize(). The former
idea was to cache all informix objects created during the planning phase
for execution later, since we rely on query statistics by using
the query descriptors. We created them in ifxGetForeignRelSize() and
reused these objects in ifxPlanForeignModify() later.

But ifxGetForeignRelSize() isn't necessarily called, so ifxPlanForeignModify()
needs to take care whether the objects are usable directly or still need
to be created.

I just copied the code from ifxGetForeignRelSize() without moving it into
a dedicated function, which fixes the problem. It might be more pretty
to have it in a dedicated function, but leave it as-is for now.

Regression tests are extended to have trigger tests as described in
issue #31.